fix: fix circular imports how Noah did in his PR
