+++ petpal/utils/image_io.py
@@ -10,13 +10,13 @@ import json
 import re
 import os
 import glob
+import pathlib
 import ants
 import nibabel
 from nibabel.filebasedimages import FileBasedHeader
 import numpy as np
 import pandas as pd
 
-from petpal.utils.bids_utils import infer_sub_ses_from_tac_path
 from . import useful_functions
 
 
@@ -598,6 +598,47 @@ def get_window_index_pairs_for_image(image_path: str, w_size: float):
     image_frame_info = get_frame_timing_info_for_nifti(image_path=image_path)
     return get_window_index_pairs_from_durations(frame_durations=image_frame_info['duration'], w_size=w_size)
 
+def infer_sub_ses_from_tac_path(tac_path: str):
+    """
+    Infers subject and session IDs from a TAC file path by analyzing the filename.
+
+    This method extracts subject and session IDs from the filename of a TAC file. It checks the
+    presence of a `sub-` and `ses-` marker in the filename, which is followed by the subject and
+    session respectively. This segment name is then formatted with each part capitalized. If no
+    subject or session is found a generic value of `UNK` is returned.
+
+    Args:
+        tac_path (str): Path of the TAC file.
+        tac_id (int): ID of the TAC.
+
+    Returns:
+        tuple: Inferred subject and session IDs.
+    """
+    path = pathlib.Path(tac_path)
+    assert path.suffix == '.tsv', '`tac_path` must point to a TSV file (*.tsv)'
+    filename = path.name
+    fileparts = filename.split("_")
+    subname = 'XXXX'
+    for part in fileparts:
+        if 'sub-' in part:
+            subname = part.split('sub-')[-1]
+            break
+    if subname == 'XXXX':
+        subname = 'UNK'
+    else:
+        name_parts = subname.split("-")
+        subname = ''.join(name_parts)
+    sesname = 'XXXX'
+    for part in fileparts:
+        if 'ses-' in part:
+            sesname = part.split('ses-')[-1]
+            break
+    if sesname == 'XXXX':
+        subname = 'UNK'
+    else:
+        name_parts = sesname.split("-")
+        sesname = ''.join(name_parts)
+    return subname, sesname
 
 def km_regional_fits_to_tsv(fit_results_dir: str, out_tsv_dir: str):
     """
